+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.2.0beta2 - 2024-10-30
+
 ### Changed
 
 - Updates to beta services Data Guard and Prompt Guard.